+++ apps/admin/src/index.tsx
@@ -17,8 +17,7 @@ import i18n from './i18n';
 import './index.css';
 import registerServiceWorker from './registerServiceWorker';
 
-// eslint-disable-next-line react-refresh/only-export-components
-function Index() {
+export function Index() {
     const {lang, loading, error} = useAppLang();
     const [i18nIsInitialized, seti18nIsInitialized] = useState(false);
