+++ libs/ui/src/components/LibraryItemsList/LibraryItemsListTable/Cell/CellInfos/CellInfos.tsx
@@ -6,14 +6,14 @@ import styled from 'styled-components';
 import {FloatingMenu, FloatingMenuAction, RecordCard} from '_ui/components';
 import useSearchReducer from '_ui/components/LibraryItemsList/hooks/useSearchReducer';
 import DeactivateRecordBtn from '_ui/components/LibraryItemsList/shared/DeactivateRecordBtn';
-import SelectCellsBtn, {SelectCellsBtnType} from '_ui/components/LibraryItemsList/shared/SelectCellsBtn';
+import SelectCellsBtn from '_ui/components/LibraryItemsList/shared/SelectCellsBtn';
 import {PreviewSize} from '_ui/constants';
 import {useSharedTranslation} from '_ui/hooks/useSharedTranslation';
 import {IRecordIdentityWhoAmI} from '_ui/types/records';
 import {ArrowsAltOutlined} from '@ant-design/icons';
 import {AntButton} from 'aristid-ds';
-import {useContext} from 'react';
-import {useEditRecordModalContext} from '../../EditRecordModalContext';
+import {useEditRecordModalContext} from '_ui/components/LibraryItemsList/LibraryItemsListTable/useEditRecordModalContext';
+import {SelectCellsBtnType} from '_ui/components/LibraryItemsList/shared/shared.utils';
 
 const Info = styled.div`
     min-width: 150px;
