+++ libs/ui/src/components/LibraryItemsList/shared/shared.utils.ts
@@ -1,0 +1,7 @@
+// Copyright LEAV Solutions 2017
+// This file is released under LGPL V3
+// License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
+export enum SelectCellsBtnType {
+    ONLY = 'ONLY',
+    ALL = 'ALL'
+}
